Cleanup event card styles
